bpf: Reorder operations in handle_ipv{4,6}
In handle_ipv{4,6}, we implement several operations:

  1. Handle NodePort services
  2. (Handle ICMPv6 messages)
  3. Swap MAC addresses
  4. Enforce host policies

However, NodePort services do not process ICMPv6 messages so we can just
process those first. In addition, it is useless to swap MAC addresses if
the packet is going to be dropped by host policies next. Thus, this
commit reorders these operations as follows.

  1. (Handle ICMPv6 messages)
  2. Handle NodePort services
  3. Enforce host policies
  4. Swap MAC addresses

Signed-off-by: Paul Chaignon <paul@cilium.io>