bpf: Fix backend selection for unconnected UDP service endpoint updates
In case of socket-lb for unconnected UDP, we select a service backend
based on the service backends count. So in case of updates, we may end
up selecting a wrong backend for existing UDP connections.

Fix:

When session affinity is disabled, we reuse the session affinity map
to store mapping for a UDP client to its selected backend using socket
cookie.
Why not enable session affinity behavior?
Session affinity is a stronger affinity in the sense that all subsequent
connections from a client will go to the same selected backend. This can
lead to uneven load-balancing.

Upgrade/Downgrade considerations:

Introduce a new flag in affinity map key using reserved bits to mark an
entry in order to indicate if its storing netns or socket cookie. In case of
downgrade, such entries won't get read as the flag wouldn't exist.

Signed-off-by: Aditi Ghag <aditi@cilium.io>